move signup and login routes to routes/index.js
