+++ views/login.ejs
@@ -5,7 +5,7 @@
   </head>
   <body>
     <!-- Beginning of Card -->
-            <form action="/users/login" method="POST">
+            <form action="/login" method="POST">
                 Username: <input type="text" name="username" ><br>
                 Password: <input type="text" name="password" ><br>
                 <input type="submit" value="Submit">
