+++ views/signup.ejs
@@ -5,7 +5,7 @@
   </head>
   <body>
     <!-- Beginning of Card -->
-            <form action="/users/signup" method="POST">
+            <form action="/signup" method="POST">
                 <input type="text" name="username" placeholder="username" required ><br>
                <input type="email" name="email" placeholder="email" required ><br>
                 <input type="text" name="password" placeholder="password" required ><br>
